requirements.txt
discord.py==1.3.4
aiohttp==3.6.2
requests==2.24.0
humanize==2.5.0
PyYAML==5.3.1
marshmallow
marshmallow-dataclass
